Remove Index action method from HomeController
The Index action method in the HomeController class has been removed. This method previously handled HTTP POST requests, populated the VillaList property of the HomeVM object with villas from the database, and set the IsAvailable property to false for villas with an even Id. The method then returned the homeVM object to the view.