Add status filtering to bookings
Modified `GetAll` in `BookingController` to accept an optional `status` parameter for filtering bookings. Updated `Index.cshtml` to read the `status` query parameter and set CSS classes for booking statuses. Adjusted the "Bookings" navigation link in `_Layout.cshtml` to include a `status` query parameter with `SD.StatusApproved`. Enhanced `loadDataTable` in `bookings.js` to accept and use the `status` parameter in the AJAX request URL.